+++ planet-mitchell/src/Eff/Mtl/Logic.hs
@@ -1,4 +1,4 @@
-module Logic
+module Eff.Mtl.Logic
   ( -- ** Logic
     Logic
   , logic