+++ planet-mitchell/src/Eff/Mtl/State.hs
@@ -1,4 +1,4 @@
-module State
+module Eff.Mtl.State
   ( -- * State
     State
   , runState